submodules/DriveX/requirements_viewcrafter.txt
av==10.0.0
decord==0.6.0
einops
imageio
imageio-ffmpeg
kornia
matplotlib
moviepy
numpy
open-clip-torch
opencv-python
Pillow
pip
pyglet==1.5.0
pytorch-lightning==1.9.3
PyYAML==6.0
roma==1.5.0
scikit-image
scikit-learn
scipy
tensorboard
timm==0.6.13
tqdm==4.65.0
transformers==4.46.2
trimesh==4.4.3
gradio==3.37.0
gradio_client==0.7.1
omegaconf==2.3.0
triton
altair==5.4.0
certifi==2024.7.4
grpcio==1.66.0
httpx==0.27.0
Pygments==2.18.0
starlette==0.38.2
xformers==0.0.21
